rework "purge" action
