+++ src/app/services/content-management.service.ts
@@ -23,12 +23,12 @@
  * along with Alfresco. If not, see <http://www.gnu.org/licenses/>.
  */
 
-import { Subject } from 'rxjs/Rx';
+import { Subject, Observable } from 'rxjs/Rx';
 import { Injectable } from '@angular/core';
 import { MatDialog } from '@angular/material';
-import { FolderDialogComponent } from '@alfresco/adf-content-services';
+import { FolderDialogComponent, ConfirmDialogComponent } from '@alfresco/adf-content-services';
 import { LibraryDialogComponent } from '../dialogs/library/library.dialog';
-import { SnackbarErrorAction } from '../store/actions';
+import { SnackbarErrorAction, SnackbarInfoAction, SnackbarAction, SnackbarWarningAction } from '../store/actions';
 import { Store } from '@ngrx/store';
 import { AppStore } from '../store/states';
 import {
@@ -38,6 +38,8 @@ import {
     SiteEntry
 } from 'alfresco-js-api';
 import { NodePermissionService } from './node-permission.service';
+import { NodeInfo, DeletedNodeInfo, DeleteStatus } from '../store/models';
+import { ContentApiService } from './content-api.service';
 
 @Injectable()
 export class ContentManagementService {
@@ -53,6 +55,7 @@ export class ContentManagementService {
 
     constructor(
         private store: Store<AppStore>,
+        private contentApi: ContentApiService,
         private permission: NodePermissionService,
         private dialogRef: MatDialog
     ) {}
@@ -143,5 +146,166 @@ export class ContentManagementService {
         return this.permission.check(sharedLink, ['update'], {
             target: 'allowableOperationsOnTarget'
         });
+    }
+
+    purgeDeletedNodes(nodes: MinimalNodeEntity[]) {
+        if (!nodes || nodes.length === 0) {
+            return;
+        }
+
+        const dialogRef = this.dialogRef.open(ConfirmDialogComponent, {
+            data: {
+                title: 'APP.DIALOGS.CONFIRM_PURGE.TITLE',
+                message: 'APP.DIALOGS.CONFIRM_PURGE.MESSAGE',
+                yesLabel: 'APP.DIALOGS.CONFIRM_PURGE.YES_LABEL',
+                noLabel: 'APP.DIALOGS.CONFIRM_PURGE.NO_LABEL'
+            },
+            minWidth: '250px'
+        });
+
+        dialogRef.afterClosed().subscribe(result => {
+            if (result === true) {
+                const nodesToDelete: NodeInfo[] = nodes.map(node => {
+                    const { name } = node.entry;
+                    const id = node.entry.nodeId || node.entry.id;
+
+                    return {
+                        id,
+                        name
+                    };
+                });
+                this.purgeNodes(nodesToDelete);
+            }
+        });
+    }
+
+    private purgeNodes(selection: NodeInfo[] = []) {
+        if (!selection.length) {
+            return;
+        }
+
+        const batch = selection.map(node => this.purgeDeletedNode(node));
+
+        Observable.forkJoin(batch).subscribe(purgedNodes => {
+            const status = this.processStatus(purgedNodes);
+
+            if (status.success.length) {
+                this.nodesPurged.next();
+            }
+            const message = this.getPurgeMessage(status);
+            if (message) {
+                this.store.dispatch(message);
+            }
+        });
+    }
+
+    private purgeDeletedNode(node: NodeInfo): Observable<DeletedNodeInfo> {
+        const { id, name } = node;
+
+        return this.contentApi
+            .purgeDeletedNode(id)
+            .map(() => ({
+                status: 1,
+                id,
+                name
+            }))
+            .catch(error => {
+                return Observable.of({
+                    status: 0,
+                    id,
+                    name
+                });
+            });
+    }
+
+    private processStatus(data: DeletedNodeInfo[] = []): DeleteStatus {
+        const status = {
+            fail: [],
+            success: [],
+            get someFailed() {
+                return !!this.fail.length;
+            },
+            get someSucceeded() {
+                return !!this.success.length;
+            },
+            get oneFailed() {
+                return this.fail.length === 1;
+            },
+            get oneSucceeded() {
+                return this.success.length === 1;
+            },
+            get allSucceeded() {
+                return this.someSucceeded && !this.someFailed;
+            },
+            get allFailed() {
+                return this.someFailed && !this.someSucceeded;
+            },
+            reset() {
+                this.fail = [];
+                this.success = [];
+            }
+        };
+
+        return data.reduce((acc, node) => {
+            if (node.status) {
+                acc.success.push(node);
+            } else {
+                acc.fail.push(node);
+            }
+
+            return acc;
+        }, status);
+    }
+
+    private getPurgeMessage(status: DeleteStatus): SnackbarAction {
+        if (status.oneSucceeded && status.someFailed && !status.oneFailed) {
+            return new SnackbarWarningAction(
+                'APP.MESSAGES.INFO.TRASH.NODES_PURGE.PARTIAL_SINGULAR',
+                {
+                    name: status.success[0].name,
+                    failed: status.fail.length
+                }
+            );
+        }
+
+        if (status.someSucceeded && !status.oneSucceeded && status.someFailed) {
+            return new SnackbarWarningAction(
+                'APP.MESSAGES.INFO.TRASH.NODES_PURGE.PARTIAL_PLURAL',
+                {
+                    number: status.success.length,
+                    failed: status.fail.length
+                }
+            );
+        }
+
+        if (status.oneSucceeded) {
+            return new SnackbarInfoAction(
+                'APP.MESSAGES.INFO.TRASH.NODES_PURGE.SINGULAR',
+                { name: status.success[0].name }
+            );
+        }
+
+        if (status.oneFailed) {
+            return new SnackbarErrorAction(
+                'APP.MESSAGES.ERRORS.TRASH.NODES_PURGE.SINGULAR',
+                { name: status.fail[0].name }
+            );
+        }
+
+        if (status.allSucceeded) {
+            return new SnackbarInfoAction(
+                'APP.MESSAGES.INFO.TRASH.NODES_PURGE.PLURAL',
+                { number: status.success.length }
+            );
+        }
+
+        if (status.allFailed) {
+            return new SnackbarErrorAction(
+                'APP.MESSAGES.ERRORS.TRASH.NODES_PURGE.PLURAL',
+                { number: status.fail.length }
+            );
+        }
+
+        return null;
     }
 }
